Refactor: Move DB queries to a separate file.
